/pinamount & /pinchannel -> /pin amount & /pin channel
